Rename tab from NAS to Authenticators
- Previously, the word NAS described the tab used to manage
  authenticator devices. Although Network Access Server (NAS) is the
  term used in RFC2865 to describe the authenticator devices, NAS
  doesn't say much for people without RADIUS experience at first
  glance. This update uses the word "Authenticator" as title which is
  also used in RFC3580 termilogy to make the tab description obvious
  for a wider audience.

- Update online documentation based on the previous change.  Here, the
  example paragraph was removed in order to keep documentation
  concise.
