+++ root/usr/share/nethesis/NethServer/Language/en/NethServer_Module_FreeRADIUS_Authenticators.php
@@ -1,6 +1,6 @@
 <?php 
 
-$L['Nas_Title'] = 'NAS'; 
+$L['Authenticators_Title'] = 'Authenticators'; 
 $L['key_label'] = 'Name';
 $L['ipaddr_label'] = 'IP address'; 
 $L['secret_label'] = 'Secret';
